# api/openapi.yaml
openapi: 3.0.3
info:
  title: LatentTone Product API
  description: |
    Phase 2+3 HTTP product surface for LatentTone (auth, sessions, stream/HLS,
    feedback, neighbor playlists, Phase 3C user playlists, embed status).

    **Try it out (Swagger / scripts):** call `POST /api/v1/auth/login` (or register),
    copy the response `token`, click **Authorize**, and paste it as a Bearer token.
    The product SPA uses the HttpOnly cookie `lt_session` only (same-origin; no
    Bearer in JavaScript). Bearer remains supported for non-browser clients.

    SPA and API share one origin (`/app/` + `/api/v1/`) — no CORS. Do not add
    `Access-Control-Allow-Origin: *`.

    Swagger UI is flag-gated (`enable_api_docs`) and is **not** a substitute for
    Phase 3 stream-smoke Gates A–D or `/dev/stream`.
  version: 0.3.1
  contact:
    name: martinsah
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: /
    description: Same origin as latenttone serve
  - url: https://latent.lt.lkeng.org
    description: Example reverse-proxied public origin (override via PUBLIC_BASE_URL)
...
paths:
...
  /api/v1/me/radio-prefs:
    get:
      tags: [radio]
      summary: Get Radio listening preferences
      description: |
        Returns per-user Radio diversification toggles. Missing rows return product
        defaults (artist cooldown, query jitter, artist penalty, bounded random, and
        Radio Bridge all **on**; `jitter_alpha` default `0.05`).
      operationId: getRadioPrefs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RadioPrefs'
        '401':
          $ref: '#/components/responses/Error'
    patch:
      tags: [radio]
      summary: Update Radio listening preferences
      description: Partial update; omitted fields keep their current (or default) values.
      operationId: patchRadioPrefs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RadioPrefsPatch'
            example:
              radio_bridge: true
              artist_cooldown: true
              jitter_alpha: 0.05
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RadioPrefs'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/me/stream-prefs:
    get:
      tags: [stream]
      summary: Get stream encode preferences
      description: |
        Per-user progressive/HLS defaults. Missing rows return `stream_format=original`
        and `bitrate_kbps=192`. `original` serves the file when the container is
        browser-safe; otherwise the server auto-transcodes to MP3. `mp3` / `aac`
        always transcode at the configured bitrate.
      operationId: getStreamPrefs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamPrefs'
        '401':
          $ref: '#/components/responses/Error'
    patch:
      tags: [stream]
      summary: Update stream encode preferences
      description: Partial update; omitted fields keep their current (or default) values.
      operationId: patchStreamPrefs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamPrefsPatch'
            example:
              stream_format: mp3
              bitrate_kbps: 192
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamPrefs'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque session id from login/register JSON field `token`.
        Preferred for Swagger Try it out: Authorize → Bearer `<token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: lt_session
      description: |
        HTTP-only cookie set on register/login. Same-origin browser clients use
        this automatically; Swagger Try-it defaults to Bearer instead.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RadioPrefs:
      type: object
      description: |
        Live Radio post-ANN diversification controls. Defaults: all boolean fields
        true; `jitter_alpha` = 0.05.
      properties:
        user_id:
          type: integer
          format: int64
        radio_bridge:
          type: boolean
          description: Every 5–7 plays inject a high-affinity bridge track and re-seed
        artist_cooldown:
          type: boolean
          description: Skip artists from the last ~3 plays when a different artist exists
        query_jitter:
          type: boolean
          description: Add small noise to the query vector before neighbor search
        artist_penalty:
          type: boolean
          description: Temporary +0.5 cosine-distance penalty for a just-played artist
        bounded_random:
          type: boolean
          description: Weighted random pick among top-N diversified candidates
        jitter_alpha:
          type: number
          format: double
          minimum: 0
          maximum: 0.5
          description: Scale of query-vector Gaussian noise when query_jitter is on
        updated_at:
          type: string
          format: date-time
      required: [user_id, radio_bridge, artist_cooldown, query_jitter, artist_penalty, bounded_random, jitter_alpha]
    RadioPrefsPatch:
      type: object
      properties:
        radio_bridge:
          type: boolean
        artist_cooldown:
          type: boolean
        query_jitter:
          type: boolean
        artist_penalty:
          type: boolean
        bounded_random:
          type: boolean
        jitter_alpha:
          type: number
          format: double
          minimum: 0
          maximum: 0.5
    StreamPrefs:
      type: object
      description: |
        Progressive/HLS encode defaults. Default `stream_format=original`,
        `bitrate_kbps=192`.
      properties:
        user_id:
          type: integer
          format: int64
        stream_format:
          type: string
          enum: [original, mp3, aac]
        bitrate_kbps:
          type: integer
          minimum: 64
          maximum: 320
        updated_at:
          type: string
          format: date-time
      required: [user_id, stream_format, bitrate_kbps]
    StreamPrefsPatch:
      type: object
      properties:
        stream_format:
          type: string
          enum: [original, mp3, aac]
        bitrate_kbps:
          type: integer
          minimum: 64
          maximum: 320
...
security:
  - bearerAuth: []